add TrackCard and SingerCard, fixed bugs
